# server/Server.Tests/_snapshots/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Backend Service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/book/{isbn}": {
      "get": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bookshelf/{bookshelfId}": {
      "get": {
        "tags": [
          "Bookshelf"
        ],
        "parameters": [
          {
            "name": "bookshelfId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerBookshelf"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bookshelf/summary/{customerId}": {
      "get": {
        "tags": [
          "Bookshelf"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookshelfSummary"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bookshelf/homeless/{customerId}": {
      "get": {
        "tags": [
          "Bookshelf"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminComment": {
        "required": [
          "adminUsername",
          "comment",
          "created"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "adminUsername": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Book": {
        "required": [
          "bookErrors",
          "isbn",
          "name",
          "pageCount",
          "subjects"
        ],
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "release": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "bookErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BookError": {
        "required": [
          "adminComment",
          "book",
          "createdAt",
          "error",
          "isbn",
          "status"
        ],
        "type": "object",
        "properties": {
          "additionalCustomerComment": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "book": {
            "$ref": "#/components/schemas/Book"
          },
          "isbn": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "error": {
            "$ref": "#/components/schemas/BookErrorType"
          },
          "adminComment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminComment"
            },
            "nullable": true,
            "readOnly": true
          },
          "status": {
            "$ref": "#/components/schemas/BookErrorStatus"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "BookErrorStatus": {
        "enum": [
          "pending",
          "closed",
          "completed"
        ],
        "type": "string"
      },
      "BookErrorType": {
        "enum": [
          "title",
          "picture",
          "author",
          "release",
          "subjects",
          "pages"
        ],
        "type": "string"
      },
...
      "BookshelfSummary": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "containsBook": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "BooktoShelf": {
        "required": [
          "book",
          "id",
          "order",
          "ranking"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "book": {
            "$ref": "#/components/schemas/Book"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "ranking": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CustomerBookshelf": {
        "required": [
          "books",
          "creationDate",
          "homelessBooks",
          "id",
          "name",
          "updatedDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BooktoShelf"
            },
            "nullable": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "homelessBooks": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme",
        "scheme": "bearer"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://dev-yf34hkokp08w5btt.uk.auth0.com/authorize?audience=/auth",
            "tokenUrl": "https://dev-yf34hkokp08w5btt.uk.auth0.com/oauth/token",
            "scopes": {}
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    },
    {
      "oauth2": []
    }
  ]
}